+++ packages/grid/x-data-grid/src/components/panel/GridPanel.tsx
@@ -8,12 +8,11 @@ import Paper from '@mui/material/Paper';
 import Popper from '@mui/material/Popper';
 import { useGridApiContext } from '../../hooks/utils/useGridApiContext';
 import { isEscapeKey } from '../../utils/keyboardUtils';
-import { GridClasses, gridClasses } from '../../constants/gridClasses';
+import { gridClasses } from '../../constants/gridClasses';
+import type { DataGridProcessedProps } from '../../models/props/DataGridProps';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
 
-type OwnerState = {
-  classes?: Partial<GridClasses>;
-};
+type OwnerState = DataGridProcessedProps;
 
 export interface GridPanelClasses {
   /** Styles applied to the root element. */
